Move ExplorationLauncher into explorer project.
